incorporated light/shading tests and a working sphere to screen

+++ src/pointlight.h
@@ -9,7 +9,7 @@
 
 struct PointLight
 {
-    glm::vec3 intensity;
+    glm::dvec3 intensity;
     glm::dvec4 position;
 };
